requirements.txt
python-dotenv==1.0.0
markdown==3.4.3
pyyaml==6.0.1
beautifulsoup4==4.12.2
rich==13.5.3
google-generativeai==0.8.4
google-genai==1.7.0
weasyprint>=60.1
pygments>=2.16.1
jinja2>=3.1.2
tiktoken
streamlit>=1.28.0
requests>=2.31.0
email-validator>=1.3.0 